- modified Viger-question-class. Since we only use it for the retrieval of "true" paths all extra parameters have been left out - fixed minor bugs

+++ src/main/java/de/lemo/dms/connectors/moodle_2_3/ExtractAndMapMoodle.java
@@ -1322,8 +1322,8 @@ private static List<Log_LMS> log_lms;
 					{						
 						if(loadedItem2.getAssignment() == Long.valueOf(loadedItem.getInfo()) && loadedItem2.getUserid().equals(loadedItem.getUserid()))// && loadedItem2.getTimemodified() == loadedItem.getTime())
 						{
-							insert.setGrade(loadedItem2.getGrade());
-							insert.setFinalgrade(loadedItem2.getGrade());
+							insert.setGrade((double)loadedItem2.getGrade());
+							insert.setFinalgrade((double)loadedItem2.getGrade());
 							break;
 						}
 					}
